requirements.txt
openai==1.65.4
langchain==0.3.20
langchain-openai==0.3.7
python-dotenv==1.0.1
langchainhub==0.1.15
pandas==2.2.2
docling==2.25.2
langchain_milvus==0.1.8
langchain-core==0.3.41
langchain-huggingface==0.1.2
langchain_docling==0.2.0
pydantic==2.10.6 